Fixed some compilation issues.
Deleted FSM Class from fsmTask(), so we can called fsmTask() in
main.cpp.

Deleted and changed startNAMETask() for NAMETask() only as it was in
Develop branch in main.cpp.

Added fsmParams into main.cpp and fsmTask.cpp:
typedef struct {
  Cow *cow;
  Fence *fence;
}fsmTaskParams;

+++ CM4/Core/Inc/cow.h
@@ -24,7 +24,7 @@ struct Acceleration {
 
 class Cow {
 public:
-    Cow(uint16_t id);
+    Cow(uint32_t id);
 
     void updatePosition(Position pos);
     void updateAcceleration(Acceleration accel);
@@ -40,7 +40,7 @@ public:
     float getDistanceToLimit() const;
 
 private:
-    uint16_t id;
+    uint32_t id;
     Position position;
     Acceleration acceleration;
     CowState state;
